+++ mkchromecast/tray_threading.py
@@ -41,9 +41,12 @@ class Worker(QObject):
         try:                        # This should fix the error socket.gaierror making the system tray to be closed.
             self.cc = casting()
             self.cc.initialize_cast()
+            self.cc.availablecc()
         except socket.gaierror:
             if debug == True:
                 print(colors.warning(':::Threading::: Socket error, CC set to 0'))
+            self.cc.availablecc == 0
+        except TypeError:
             self.cc.availablecc == 0
         if len(self.cc.availablecc) == 0 and tray == True:
             availablecc = []