+++ changelog.md
@@ -1,6 +1,6 @@
-* mkchromecast (0.3.7) **released**: 2016/12/30
+* mkchromecast (0.3.7) **unreleased**
 
-    - node streaming updated to work with `node v7.2.1`.
+    - node streaming updated to work with `node v7.8.0`.
     - Added ALSA device configuration in preferences pane.
     - Improved systray stability.
     - gstreamer support for capturing audio.
@@ -15,6 +15,13 @@
       Closes [#32](https://github.com/muammar/mkchromecast/pull/32).
     - Making ping code python 3 compatible. Closes:
       [#35](https://github.com/muammar/mkchromecast/pull/35).
+    - Fixed problem when having various Google cast devices. Closes
+      [#50](https://github.com/muammar/mkchromecast/issues/50)
+    - Added support to 192000Hz sampling rate support (Closes:
+      [#39](https://github.com/muammar/mkchromecast/issues/39)).
+    - Fixed a minimal problem with size of preferences pane.
+    - Video support.
+    - node is a supported backend for streaming video.
 
 * mkchromecast (0.3.6) **released**: 2016/09/19